+++ CHANGELOG.md
@@ -630,6 +630,7 @@
   - /zone_delete -> /zone_end
 * Added `background_tod` as an area parameter. Subparameters defined within it will indicate compatible clients to switch to a given background according to the current time of day present in the client's area.
 * Added /files_area (command alias /fa), which returns all visible players in the area who have set their files.
+* Added command alias /l for /look.
 * Fixed scream_range in area list yaml files not supporting the keyword <ALL> to indicate all areas should be able to receive a scream coming from a particular area
 * Fixed scream_range in area list yaml files not checking if the areas a scream can reach to from a particular area exist
 * Fixed /scream, /whisper and /party_whisper raising errors if a message was sent to a deafened player with a bypass message starter. They now sent messages but filtered